Refactor unit tests for s3 driver
Reduced test duplications for single store and multistore cases.
